feat: MatchCreation and list enchanced
